refactor(era5): new module list_metrics
This allows easier import of metrics list into era5 submodules
and avoids circular imports.
